+++ packages/svelte/svelte.config.js
@@ -3,22 +3,10 @@ import preprocess from 'svelte-preprocess'
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-  // Consult https://github.com/sveltejs/svelte-preprocess
-  // for more information about preprocessors
   preprocess: preprocess(),
 
   kit: {
     adapter: adapter(),
-    // vite: () => ({
-    //   server: {
-    //     fs: {
-    //       allow: ['./package'],
-    //     },
-    //   },
-    // }),
-    package: {
-      exports: (filepath) => !filepath.endsWith('.svelte'),
-    },
   },
 }