+++ scripts/lib/package.ts
@@ -3,14 +3,30 @@ import path from 'path'
 
 const VUE_PACKAGE_PATH = path.join(__dirname, '../../packages/vue')
 
-const copyPackageFile = async () => {
+const createExports = async () => {
+  const fillComponents = await fs.readdir(`${VUE_PACKAGE_PATH}/src/lib/fill`)
+  const strokeComponents = await fs.readdir(`${VUE_PACKAGE_PATH}/src/lib/stroke`)
+
+  const exports: Record<string, string> = {
+    './package.json': './package.json',
+  }
+
+  fillComponents.forEach((componentName) => {
+    exports[`./${componentName}`] = `./fill/${componentName}`
+  })
+  strokeComponents.forEach((componentName) => {
+    exports[`./${componentName}`] = `./stroke/${componentName}`
+  })
+
+  return exports
+}
+
+const copyPackageFile = async (exports: Record<string, string>) => {
   const packageStr = await fs.readFile(`${VUE_PACKAGE_PATH}/package.json`, { encoding: 'utf-8' })
   const _package = JSON.parse(packageStr)
   const newPackage = {
     name: _package.name,
     version: _package.version,
-    types: './index.d.ts',
-    main: './index.js',
     devDependencies: _package.devDependencies,
     keywords: _package.keywords,
     author: _package.author,
@@ -18,18 +34,9 @@ const copyPackageFile = async () => {
     repository: _package.repository,
     license: _package.license,
     sideEffects: _package.sideEffects,
-    exports: {
-      './package.json': './package.json',
-      '.': './index.js',
-    },
+    exports,
   }
   await fs.writeFile(`${VUE_PACKAGE_PATH}/package/package.json`, JSON.stringify(newPackage, null, 2))
-}
-
-const copyIndexFile = async () => {
-  const _exports = await fs.readFile(`${VUE_PACKAGE_PATH}/src/lib/index.ts`)
-  await fs.writeFile(`${VUE_PACKAGE_PATH}/package/index.d.ts`, _exports)
-  await fs.writeFile(`${VUE_PACKAGE_PATH}/package/index.js`, _exports)
 }
 
 (async () => {
@@ -39,8 +46,9 @@ const copyIndexFile = async () => {
     console.log('Can not find \'vue/package\' directory, skipping to next steps')
   }
   await fs.mkdir(`${VUE_PACKAGE_PATH}/package`)
-  await copyPackageFile()
-  await copyIndexFile()
+
+  const exports = await createExports()
+  await copyPackageFile(exports)
 
   await fs.mkdir(`${VUE_PACKAGE_PATH}/package/fill`)
   await fs.mkdir(`${VUE_PACKAGE_PATH}/package/stroke`)
@@ -48,4 +56,5 @@ const copyIndexFile = async () => {
   await fs.cp(`${VUE_PACKAGE_PATH}/src/lib/fill`, `${VUE_PACKAGE_PATH}/package/fill`, { recursive: true })
   await fs.cp(`${VUE_PACKAGE_PATH}/src/lib/stroke`, `${VUE_PACKAGE_PATH}/package/stroke`, { recursive: true })
   await fs.cp(`${VUE_PACKAGE_PATH}/LICENSE`, `${VUE_PACKAGE_PATH}/package/LICENSE`)
+  await fs.cp(`${VUE_PACKAGE_PATH}/README.md`, `${VUE_PACKAGE_PATH}/package/README.md`)
 })()