+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 
 All notable changes to this project Sveegy will be documented in this file.
 
+## @sveegy/icons-vue [0.1.0] - 03-18-2022
+
+- Added README to package
+- Changed exports
+```js
+// not working
+import { ... } from '@sveegy/icons-svelte'
+
+// working
+import ... from '@sveegy/icons-svelte/SvAddAssetsFill.svelte'
+```
+
+## @sveegy/icons-svelte [0.3.0] - 03-18-2022
+
+- Changed exports
+```js
+// not working
+import { ... } from '@sveegy/icons-svelte'
+
+// working
+import ... from '@sveegy/icons-svelte/SvAddAssetsFill.svelte'
+```
+
+## @sveegy/icons-vue [0.0.3] - 03-14-2022
+
+- Added README.md to package
+
 ## @sveegy/icons-svelte [0.3.1] - 02-24-2022
 
 - Added new icons for both types: cloud, cloud-download, cloud-upload